+++ packages/controls/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add WP 6.8 compatibility.
 - Improve overall codes.
 - Improve Unit Inputs for 'No Unit' state. 
+- Border feature improved to show a warning for color if width was set and the color wasn't!
 
 
 ## 1.1.2 (2025-03-29)